+++ com/hbm/entity/effect/EntityEMPBlast.java
@@ -3,7 +3,6 @@ package com.hbm.entity.effect;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.effect.EntityLightningBolt;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.world.World;
 
@@ -24,7 +23,7 @@ public class EntityEMPBlast extends Entity {
 
 	@Override
 	protected void entityInit() {
-		this.dataWatcher.addObject(16, Integer.valueOf((int) 0));
+		this.dataWatcher.addObject(16, Integer.valueOf(0));
 	}
 
     @Override
